Added all the required files to operate the VRBox
We can run this project to view Assets/eso0932.jpg in the form of a 3d virtual reality perspective.
Running the Game will allow us to view the image in a 3 dimensional spherical form, such that changing the field of view can allow the user to see different parts of the image, as if they were looking at the sky.
This project was done as a Part of my Audit Course.